requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.4
sentence-transformers>=3.0.1
pymupdf>=1.24.7
pydantic>=2.7
python-dotenv>=1.0.1
streamlit>=1.36
psutil>=5.9
langchain-ollama>=0.1.0
numpy
pandas
langchain-huggingface>=0.1.0
datasets>=2.19.0
ragas>=0.2.2
langchain-ollama>=0.1.0
langchain-community>=0.3.0
